% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stage2.R
\name{BLOCKPCAs}
\alias{BLOCKPCAs}
\title{estimates the pricipal components and corresponding eigen values for each cluster.}
\usage{
BLOCKPCAs(corrmat, set, num)
}
\arguments{
\item{corrmat}{correlation matrix}

\item{set}{the clustering indicator}

\item{num}{The number of principal components}
}
\value{
returns the block pcs and corresponding eigenvalues.
}
\description{
estimates the pricipal components and corresponding eigen values for each cluster.
}
