% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stage2.R
\name{ComputeCluster}
\alias{ComputeCluster}
\title{Hierarchical Clustering based on Correlation distance}
\usage{
ComputeCluster(corrmat, cutoff, groups, hclustopt = "ward.D")
}
\arguments{
\item{corrmat}{correlation matrix}

\item{cutoff}{thresholding value}

\item{groups}{the number of the groups to be cluster}

\item{hclustopt}{the hierarchical cluster option (ex: "cen", "ward.D" )}
}
\value{
matrix containing the features in the 1st row and the clustering indicator corresponding to the feature in the 2nd row
}
\description{
Use the covariance thresholding method introduced by Bickel and Levina (2008) and the hierarchical clustering
method to partition the features into the multiple disjoint clusters.
}
