% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stage1.R
\name{fit_stage1}
\alias{fit_stage1}
\title{Stage (I) estimation}
\usage{
fit_stage1(x, y)
}
\arguments{
\item{x}{is matrix n times p containing  all candidate features (p) in columns and all observations (n) in rows.}

\item{y}{is outcome variable.}
}
\value{
provides the HSIC statistic as between marginal feature and the outcome as two-variable independence test
 in Reproducing Kernel Hilbert Spaces (RKHS).
}
\description{
Use a sure Independence  Screening (SIS) procedure based on a Hilbert-Schmidt Independence Criterion (HSIC) in the reproducing Hilbert kernel space to select
 a set of important features.
}
\examples{
x=matrix( runif(100,0,1), ncol=10)
y=matrix( runif(10,0,1), ncol=1)
fit_stage1( x, y)
}
